refactor : Seperate modal component
